Add image gallery page with folder grid, lightbox, and job creation
Extract CreateJobDialog to shared component with initialStartingImage prop.
Build ImageRepo page: folder grid with thumbnails, image grid with hover
delete, lightbox modal, and "Use as Starting Image" flow.
